// wow-minmax-ui/src/app/model/equipment/ItemSlot.ts
export enum ItemSlot {
	HEAD = 'HEAD',
	NECK = 'NECK',
	SHOULDER = 'SHOULDER',
	BACK = 'BACK',
	CHEST = 'CHEST',
	WRIST = 'WRIST',
	HANDS = 'HANDS',
	WAIST = 'WAIST',
	LEGS = 'LEGS',
	FEET = 'FEET',
	FINGER_1 = 'FINGER_1',
	FINGER_2 = 'FINGER_2',
	TRINKET_1 = 'TRINKET_1',
	TRINKET_2 = 'TRINKET_2',
	MAIN_HAND = 'MAIN_HAND',
	OFF_HAND = 'OFF_HAND',
	RANGED = 'RANGED',
}